From rstory 2018-10-02: "patch-v3-header-asn-double-count.diff"
Here's a mini-patch to fix a double-count of sizes for 2 asn sequences
in v3 header...

also, when you are calling usm_calc_offsets you pass 0 for engine boots
and engine time. This causes the lengths to be the minimal size and the
actual lengths are most likely to be longer.

I also noticed that it has hardcoded values for auth/priv parameter
lengths, which is unlikely to work for larger sizes. I'm still working
on a patch for that.

Also, I never heard back from you on my last message (forward encoding
not working with the patch you sent me applied to 5.8 or 5-8-patches).
Maybe you have some other change checked in to your local branch that
makes things work?